[#3] feature :  필터 컨테이너 분리 완료
필터 적용 버튼 눌렀을 때 필터 아이템이 적용되도록 수정
