+++ src/pages/TravleListPage.tsx
@@ -1,23 +1,5 @@
-import { AddIcon, MinusIcon } from '@chakra-ui/icons';
-import {
-  Box,
-  Button,
-  Collapse,
-  Container,
-  Divider,
-  Grid,
-  Heading,
-  RangeSlider,
-  RangeSliderFilledTrack,
-  RangeSliderMark,
-  RangeSliderThumb,
-  RangeSliderTrack,
-  Tag,
-  TagLabel,
-  TagLeftIcon,
-  Text,
-  useDisclosure,
-} from '@chakra-ui/react';
+import { Box, Button, Collapse, Container, Grid, Text, useDisclosure } from '@chakra-ui/react';
+import FilterContainer from 'components/FilterContainer';
 import MainLayout from 'components/MainLayout';
 import TravleContent from 'components/TravleContent';
 import { useState } from 'react';
@@ -25,37 +7,48 @@ import { useState } from 'react';
 import { useLoaderData } from 'react-router-dom';
 import { travleContent } from 'types';
 
+export interface FilterItems {
+  price: number[];
+  spaceCategory: string[];
+}
+
 const Main = () => {
   const data = useLoaderData() as travleContent[];
   const spaceCategoryData = Array.from(
     data.reduce((arr, currData) => arr.add(currData.spaceCategory), new Set<string>()),
   );
-
   const { isOpen, onToggle } = useDisclosure();
-  const [sliderValue, setSliderValue] = useState([0, 100]);
-  const isFilterPrice = sliderValue[0] !== 0 || sliderValue[1] !== 100;
 
-  const [spaceCategoriesFilter, setSpaceCatoriesFilter] = useState<string[]>(spaceCategoryData);
-  const isSpaceCategoriesFilter = spaceCategoriesFilter.length !== 0;
+  const [filterItems, setFilterItems] = useState<FilterItems>({
+    price: [0, 100],
+    spaceCategory: [],
+  });
+
+  const setItems = (items: FilterItems) => {
+    setFilterItems(items);
+  };
+
   const getFilteredData = () => {
-    debugger;
-    if (isFilterPrice && isSpaceCategoriesFilter) {
+    const { price, spaceCategory } = filterItems;
+    const isPriceFilter = price[0] !== 0 && price[1] !== 100;
+    const isSpaceCategoryFilter = spaceCategory.length > 0;
+    if (isPriceFilter && isSpaceCategoryFilter) {
       return data.filter(content => {
-        const price = content.price / 1000;
-        return sliderValue[0] <= price &&
-          price <= sliderValue[1] &&
-          spaceCategoriesFilter.includes(content.spaceCategory)
+        const contentPrice = content.price / 1000;
+        return price[0] <= contentPrice &&
+          contentPrice <= price[1] &&
+          spaceCategory.includes(content.spaceCategory)
           ? content
           : null;
       });
-    } else if (isFilterPrice) {
+    } else if (isPriceFilter) {
       return data.filter(content => {
-        const price = content.price / 1000;
-        return sliderValue[0] <= price && price <= sliderValue[1] ? content : null;
+        const contentPrice = content.price / 1000;
+        return price[0] <= contentPrice && contentPrice <= price[1] ? content : null;
       });
-    } else if (isSpaceCategoriesFilter) {
+    } else if (isSpaceCategoryFilter) {
       return data.filter(content =>
-        spaceCategoriesFilter.includes(content.spaceCategory) ? content : null,
+        spaceCategory.includes(content.spaceCategory) ? content : null,
       );
     }
     return data;
@@ -63,101 +56,16 @@ const Main = () => {
 
   const travelContentsView = getFilteredData();
 
-  const addCategory = (space: string) => {
-    !spaceCategoriesFilter.includes(space) &&
-      setSpaceCatoriesFilter([...spaceCategoriesFilter, space]);
-  };
-
-  const removeCategory = (space: string) => {
-    setSpaceCatoriesFilter(spaceCategoriesFilter.filter(data => data !== space));
-  };
-
   return (
     <MainLayout>
       <Container as='section'>
         <Button onClick={onToggle}>필터</Button>
         <Collapse in={isOpen} animateOpacity>
-          <Box padding={5}>
-            <Heading fontSize={'xl'}>가격</Heading>
-            <Box minH='60px' padding='10x 5px'>
-              <RangeSlider
-                maxW='90%'
-                aria-label={['min', 'max']}
-                defaultValue={sliderValue}
-                onChange={val => setSliderValue(val)}
-              >
-                <RangeSliderMark value={0} mt='1' fontSize='sm'>
-                  0
-                </RangeSliderMark>
-                <RangeSliderMark value={50} mt='1' fontSize='sm'>
-                  50,000
-                </RangeSliderMark>
-                <RangeSliderMark value={100} mt='1' fontSize='sm'>
-                  100,000
-                </RangeSliderMark>
-                <RangeSliderMark
-                  value={sliderValue[0]}
-                  textAlign='center'
-                  color='gray.500'
-                  mt='-10'
-                  ml='-5'
-                  w='15'
-                >
-                  {(sliderValue[0] * 1000).toLocaleString()}
-                </RangeSliderMark>
-                <RangeSliderMark
-                  value={sliderValue[1]}
-                  textAlign='center'
-                  color='gray.500'
-                  mt='-10'
-                  ml='-5'
-                  w='15'
-                >
-                  {(sliderValue[1] * 1000).toLocaleString()}
-                </RangeSliderMark>
-                <RangeSliderTrack>
-                  <RangeSliderFilledTrack />
-                </RangeSliderTrack>
-                <RangeSliderThumb boxSize={6} index={0} />
-                <RangeSliderThumb boxSize={6} index={1} />
-              </RangeSlider>
-            </Box>
-          </Box>
-          <Divider />
-          <Box padding={5} display='flex' flexDirection='column'>
-            <Heading fontSize={'xl'}>장소</Heading>
-            <Box display='flex' gap={3} padding='10px 5px'>
-              {spaceCategoryData.map(space =>
-                spaceCategoriesFilter.includes(space) ? null : (
-                  <Tag
-                    size='md'
-                    variant='subtle'
-                    colorScheme='cyan'
-                    key={space}
-                    onClick={() => addCategory(space)}
-                  >
-                    <TagLeftIcon boxSize='12px' as={AddIcon} />
-                    <TagLabel>{space}</TagLabel>
-                  </Tag>
-                ),
-              )}
-            </Box>
-            <Divider />
-            <Box display='flex' gap={3} padding='10px 5px'>
-              {spaceCategoriesFilter.map(space => (
-                <Tag
-                  size='md'
-                  variant='subtle'
-                  colorScheme='teal'
-                  key={space}
-                  onClick={() => removeCategory(space)}
-                >
-                  <TagLeftIcon boxSize='12px' as={MinusIcon} />
-                  <TagLabel>{space}</TagLabel>
-                </Tag>
-              ))}
-            </Box>
-          </Box>
+          <FilterContainer
+            defaultSpaceCategory={spaceCategoryData}
+            filterItems={filterItems}
+            setFilterItems={setItems}
+          />
         </Collapse>
       </Container>
       <Box as='section'>
